Renamed core to not confuse the reader
