+++ BenchmarkProblems/IsingSpinGlassProblem.py
@@ -4,9 +4,9 @@ from typing import TypeAlias
 import numpy as np
 
 from BenchmarkProblems.BenchmarkProblem import BenchmarkProblem
-from Core.FullSolution import FullSolution
-from Core.PS import PS, STAR
-from Core.SearchSpace import SearchSpace
+from FirstPaper.FullSolution import FullSolution
+from FirstPaper.PS import PS, STAR
+from FirstPaper.SearchSpace import SearchSpace
 
 LinkValue: TypeAlias = int
