+++ FirstPaper/ArchivePSMiner.py
@@ -7,18 +7,18 @@ import numpy as np
 
 import utils
 from BenchmarkProblems.BenchmarkProblem import BenchmarkProblem
-from Core.EvaluatedPS import EvaluatedPS
-from Core.PRef import PRef
-from Core.PS import PS
-from Core.PSMetric.Atomicity import Atomicity
-from Core.PSMetric.MeanFitness import MeanFitness
-from Core.PSMetric.Metric import Metric
-from Core.PSMetric.Simplicity import Simplicity
-from Core.SearchSpace import SearchSpace
-from Core.TerminationCriteria import TerminationCriteria, PSEvaluationLimit, IterationLimit
-from Core.get_init import just_empty
-from Core.get_local import specialisations
-from Core.selection import truncation_selection
+from FirstPaper.EvaluatedPS import EvaluatedPS
+from FirstPaper.PRef import PRef
+from FirstPaper.PS import PS
+from FirstPaper.PSMetric.Atomicity import Atomicity
+from FirstPaper.PSMetric.MeanFitness import MeanFitness
+from FirstPaper.PSMetric.Metric import Metric
+from FirstPaper.PSMetric.Simplicity import Simplicity
+from FirstPaper.SearchSpace import SearchSpace
+from FirstPaper.TerminationCriteria import TerminationCriteria, PSEvaluationLimit, IterationLimit
+from FirstPaper.get_init import just_empty
+from FirstPaper.get_local import specialisations
+from FirstPaper.selection import truncation_selection
 from FSStochasticSearch.GA import GA
 from FSStochasticSearch.Operators import SinglePointFSMutation, TwoPointFSCrossover, TournamentSelection
 from PSMiners.AbstractPSMiner import AbstractPSMiner
@@ -30,7 +30,7 @@ SelectionType: TypeAlias = [[list[EvaluatedPS], int], list[EvaluatedPS]]
 
 
 class ArchivePSMiner(AbstractPSMiner):
-    """This class is the Core miner, which outputs a Core catalog when used right"""
+    """This class is the FirstPaper miner, which outputs a FirstPaper catalog when used right"""
     """There are many parts that can be modified, and these were tested in the paper, 
     but you should probably just use with_default_settings as a constructor"""
