+++ FirstPaper/PSMetric/Simplicity.py
@@ -1,8 +1,8 @@
 import numpy as np
 
-from Core.PRef import PRef
-from Core.PS import PS, STAR
-from Core.PSMetric.Metric import Metric
+from FirstPaper.PRef import PRef
+from FirstPaper.PS import PS, STAR
+from FirstPaper.PSMetric.Metric import Metric
 
 
 class Simplicity(Metric):
